event / translation

+++ src/Core/Infrastructure/Event/Doctrine/RemoveEntityEvent.php
@@ -8,6 +8,8 @@ use App\Core\Domain\Entity\Grave;
 use App\Core\Domain\EventListener\Doctrine\PostRemoveListener;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\Event\LifecycleEventArgs;
+use Exception;
+use Symfony\Component\Filesystem\Exception\FileNotFoundException;
 use Symfony\Component\Filesystem\Filesystem;
 
 class RemoveEntityEvent extends PostRemoveListener
@@ -24,37 +26,43 @@ class RemoveEntityEvent extends PostRemoveListener
     public function postRemove(LifecycleEventArgs $args): void
     {
         $entity = $args->getObject();
-        $this->flash->handleNotification($entity);
 
         // remove orphans images
         if ($entity instanceof Grave) {
             $images = $entity->getImages();
 
-            $filesystem = new Filesystem();
-            $paths = [];
-            $thumbPaths = [];
+            try {
+                if ($images->count()) {
+                    $filesystem = new Filesystem();
+                    $paths = [];
+                    $thumbPaths = [];
 
-            /** @var FileGrave $image */
-            foreach ($images as $image) {
-                $paths[] = $this->targetDirectory . '/' .
-                    $image->getName() . '.'. $image->getExtension()->value;
-                $thumbPaths[] = $this->targetThumbnailDirectory . '/' .
-                    $image->getName() . '.'. $image->getExtension()->value;
-                $this->em->remove($image);
+                    /** @var FileGrave $image */
+                    foreach ($images as $image) {
+                        $paths[] = $this->targetDirectory . '/' .
+                            $image->getName() . '.'. $image->getExtension()->value;
+                        $thumbPaths[] = $this->targetThumbnailDirectory . '/' .
+                            $image->getName() . '.'. $image->getExtension()->value;
+                        $this->em->remove($image);
+                    }
+
+                    foreach ($paths as $path) {
+                        if ($filesystem->exists($path)) {
+                            $filesystem->remove($path);
+                        }
+                    }
+
+                    foreach ($thumbPaths as $thumbPath) {
+                        if ($filesystem->exists($thumbPath)) {
+                            $filesystem->remove($thumbPath);
+                        }
+                    }
+                }
+            } catch (Exception $e) {
+                throw new FileNotFoundException($e->getMessage());
             }
             $this->em->flush();
-
-            foreach ($paths as $path) {
-                if ($filesystem->exists($path)) {
-                    $filesystem->remove($path);
-                }
-            }
-
-            foreach ($thumbPaths as $thumbPath) {
-                if ($filesystem->exists($thumbPath)) {
-                    $filesystem->remove($thumbPath);
-                }
-            }
+            $this->flash->handleNotification($entity);
         }
     }
 }
